refactor(cli): split the audit command into per-concern modules
`cli_args/audit.rs` held four unrelated jobs in one file: the command
itself, the registry wire format, the dependency graph sent for audit,
and everything the `--fix` flag does. Split along those lines:

- `report` — the audit response and the report derived from it
- `request` — turning a lockfile into the graph the registry audits
- `paths` — the dependency paths reported with each advisory
- `render` — the JSON and text renderings
- `version_ranges` — the semver questions asked of advisory ranges
- `fix` — overrides, ignores, and dependency updates

`audit.rs` keeps the argument types and the command body, at 2126 lines
down to 624. Re-exports are explicit rather than glob, and cover only
what is used outside the owning module; the tests now import each item
from the module that defines it.

Pure code motion: no behaviour change, verified by comparing every
content line of the original against the union of the new modules.
